Moved Redux utils to reduxlike_utils.js
